+++ LuckyoneUI/Profiles/ElvUI.lua
@@ -298,14 +298,15 @@ function Private:Setup_NamePlates(installer)
 
 	-- NamePlates general
 	E.db.nameplates.classColorNames = true
+	E.db.nameplates.clickSize.height = 22
 	E.db.nameplates.clickSize.width = 210
 	E.db.nameplates.fadeIn = false
 	E.db.nameplates.lowHealthThreshold = 0
 	E.db.nameplates.overlapH = 1.2
-	E.db.nameplates.overlapV = 1.6
+	E.db.nameplates.overlapV = 1.8
 	E.db.nameplates.statusbar = Private.Texture
-	E.db.nameplates.threat.useSoloColor = true
 	E.db.nameplates.threat.skipGoodColor = true
+	E.db.nameplates.threat.useSoloColor = true
 	E.db.nameplates.useBlizzardAuras = Private.isRetail -- Use filter prio in nonRetail
 
 	-- NamePlates misc